hspace: change handling of N+1 bad clusters
Currently we just print a fake result and exit early. This is bad, since
it doesn't use the same codepaths for all the result printing, and has
already led to a bug where hspace looks like completely ignoring the
tiered specs request.

This patch changes the code to override the computation, instead of the
printout, thus allowing tiered specs results for such cases.