Remove the noLimit values and always use limits
This patch moves from allowing no-limits for disk/cpu ratios, and always
use a real limit. For disk, it's simple since we use 0, which means no
reservations for disks. For CPU, we set an (arbitrary) limit of 64 v/p,
which should be reasonable as a default limit (it can be changed via the
command line).